requirements.txt
aniso8601==9.0.1
astroid==2.3.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
dill==0.3.8
Faker==24.11.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
idna==3.7
iniconfig==2.0.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.3
lazy-object-proxy==1.4.3
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mccabe==0.6.1
more-itertools==10.2.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
py==1.11.0
PyJWT==2.8.0
pylint==2.4.4
PyMySQL==1.1.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-env==0.6.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
tomlkit==0.12.4
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.11.2
